# Backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.CreateWorkOrderRequest:
    properties:
      assigned_to:
        example: 1
        type: integer
      customer_name:
        example: John Doe
        type: string
      description:
        example: Production of 100 jerseys
        type: string
      end_date:
        example: "2024-03-26T21:41:49Z"
        type: string
      estimated_cost:
        example: 1000
        type: number
      items:
        items:
          $ref: '#/definitions/handlers.WorkOrderItemInput'
        type: array
      order_id:
        example: 1
        type: integer
      start_date:
        example: "2024-03-24T21:41:49Z"
        type: string
      tasks:
        items:
          $ref: '#/definitions/handlers.WorkOrderTaskInput'
        type: array
      work_type:
        example: production
        type: string
    required:
    - assigned_to
    - customer_name
    - description
    - end_date
    - estimated_cost
    - order_id
    - start_date
    - work_type
    type: object
...
  handlers.ErrorResponse:
    description: Standard error response model
    properties:
      error:
        example: Invalid request parameters
        type: string
    type: object
...
  handlers.SuccessResponse:
    description: Standard success response model
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
...
  handlers.UpdateWorkOrderRequest:
    properties:
      assigned_to:
        example: 1
        type: integer
      customer_name:
        example: John Doe
        type: string
      description:
        example: Production of 100 jerseys
        type: string
      end_date:
        example: "2024-03-26T21:41:49Z"
        type: string
      estimated_cost:
        example: 1000
        type: number
      items:
        items:
          $ref: '#/definitions/handlers.WorkOrderItemInput'
        type: array
      start_date:
        example: "2024-03-24T21:41:49Z"
        type: string
      tasks:
        items:
          $ref: '#/definitions/handlers.WorkOrderTaskInput'
        type: array
      work_type:
        example: production
        type: string
    required:
    - assigned_to
    - customer_name
    - description
    - end_date
    - estimated_cost
    - start_date
    - work_type
    type: object
...
  handlers.WorkOrderItemInfo:
    properties:
      description:
        example: Fabric material
        type: string
      id:
        example: 1
        type: integer
      item_id:
        example: 1
        type: integer
      quantity:
        example: 100
        type: integer
      total_price:
        example: 1000
        type: number
      unit_price:
        example: 10
        type: number
    type: object
  handlers.WorkOrderItemInput:
    properties:
      description:
        example: Fabric material
        type: string
      item_id:
        example: 1
        type: integer
      quantity:
        example: 100
        type: integer
      unit_price:
        example: 10
        type: number
    required:
    - item_id
    - quantity
    - unit_price
    type: object
  handlers.WorkOrderResponse:
    properties:
      actual_cost:
        example: 950
        type: number
      assigned_to:
        example: 1
        type: integer
      completion_notes:
        example: Completed on time
        type: string
      created_at:
        example: "2024-03-24T21:41:49Z"
        type: string
      created_by:
        example: admin
        type: string
      customer_name:
        example: John Doe
        type: string
      description:
        example: Production of 100 jerseys
        type: string
      end_date:
        example: "2024-03-26T21:41:49Z"
        type: string
      estimated_cost:
        example: 1000
        type: number
      id:
        example: 1
        type: integer
      items:
        items:
          $ref: '#/definitions/handlers.WorkOrderItemInfo'
        type: array
      order_id:
        example: 1
        type: integer
      spk_number:
        example: SPK-2024-001
        type: string
      start_date:
        example: "2024-03-24T21:41:49Z"
        type: string
      status:
        example: draft
        type: string
      tasks:
        items:
          $ref: '#/definitions/handlers.WorkOrderTaskItem'
        type: array
      updated_at:
        example: "2024-03-24T21:41:49Z"
        type: string
      updated_by:
        example: admin
        type: string
      work_type:
        example: production
        type: string
    type: object
  handlers.WorkOrderTaskInput:
    properties:
      assigned_to:
        example: 1
        type: integer
      description:
        example: Cut the fabric according to patterns
        type: string
      end_date:
        example: "2024-03-24T23:41:49Z"
        type: string
      start_date:
        example: "2024-03-24T21:41:49Z"
        type: string
      task_name:
        example: Cutting
        type: string
    required:
    - assigned_to
    - end_date
    - start_date
    - task_name
    type: object
  handlers.WorkOrderTaskItem:
    properties:
      assigned_to:
        example: 1
        type: integer
      description:
        example: Cut the fabric according to patterns
        type: string
      end_date:
        example: "2024-03-24T23:41:49Z"
        type: string
      id:
        example: 1
        type: integer
      notes:
        example: Ready for cutting
        type: string
      start_date:
        example: "2024-03-24T21:41:49Z"
        type: string
      status:
        example: pending
        type: string
      task_name:
        example: Cutting
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Vomo backend API documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Vomo API
  version: "1.0"
paths:
...
  /work-orders:
    get:
      description: Get all work orders
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.WorkOrderResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all work orders
      tags:
      - WorkOrder
    post:
      consumes:
      - application/json
      description: Create a new work order with the provided details
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Work Order Data
        in: body
        name: workOrder
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateWorkOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.WorkOrderResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new work order
      tags:
      - WorkOrder
  /work-orders/{id}:
    delete:
      description: Delete an existing work order
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Work Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Work order not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a work order
      tags:
      - WorkOrder
    get:
      description: Get work order details by ID
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Work Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.WorkOrderResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Work order not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a work order by ID
      tags:
      - WorkOrder
    put:
      consumes:
      - application/json
      description: Update an existing work order with new details
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Work Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Work Order Data
        in: body
        name: workOrder
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateWorkOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.WorkOrderResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Work order not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a work order
      tags:
      - WorkOrder
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"